[refactor]: OAuth2관련 컴포넌트 page.tsx -> components폴더로 이동
